mcs add tests for groups and mock same function twice
